[FIX] web_editor: allow deleting all images in table cells
Before this commit, it was impossible to delete an image which was a
direct child of a table cell using the trash-icon button in the
toolbar.

This commit fixes the behavior by changing the logic quite a lot, so
it is kinda tricky (see https://github.com/odoo/odoo/commit/ca65fae6aa8b2ab803b35f4d19658aff73542118
for more explaination).

Closes https://github.com/odoo/odoo/issues/22523